Rename e2e step and method
We are rewording some text which was previously referenced in method names.
Updating these to be more generic.
