+++ e2e-tests/tests/group-placements/02_update_appointment_enforcement.spec.ts
@@ -2,7 +2,7 @@ import test from '../../fixtures/test'
 import signIn from '../../steps/signIn'
 import searchForASession from '../../steps/searchForASession'
 import selectASession from '../../steps/selectASession'
-import clickUpdateAnAppointment from '../../steps/clickUpdateAnAppointment'
+import viewAppointmentFromList from '../../steps/viewAppointmentFromList'
 import completeCheckAppointmentDetails from '../../steps/completeCheckAppointmentDetails'
 import ConfirmPage from '../../pages/appointments/confirmPage'
 import { checkAppointmentOnDelius } from '../../steps/delius'
@@ -26,7 +26,7 @@ test('Update a session appointment with an enforceable outcome', async ({
 
   await sessionPage.expect.toSeeAppointments()
 
-  const checkAppointmentDetailsPage = await clickUpdateAnAppointment(page, sessionPage, personOnProbation.crn)
+  const checkAppointmentDetailsPage = await viewAppointmentFromList(page, sessionPage, personOnProbation.crn)
   const chooseSupervisorPage = await completeCheckAppointmentDetails(page, checkAppointmentDetailsPage)
 
   const attendanceOutcomePage = await completeChooseSupervisor(page, chooseSupervisorPage, team)
