+++ e2e-tests/tests/group-placements/03_update_appointment_attended_enforcement.spec.ts
@@ -2,7 +2,7 @@ import test from '../../fixtures/test'
 import signIn from '../../steps/signIn'
 import searchForASession from '../../steps/searchForASession'
 import selectASession from '../../steps/selectASession'
-import clickUpdateAnAppointment from '../../steps/clickUpdateAnAppointment'
+import viewAppointmentFromList from '../../steps/viewAppointmentFromList'
 import completeCheckAppointmentDetails from '../../steps/completeCheckAppointmentDetails'
 import { completeAttendedEnforceableOutcome } from '../../steps/completeAttendanceOutcome'
 import completeCompliance from '../../steps/completeCompliance'
@@ -27,7 +27,7 @@ test('Update a session appointment with an attended but enforceable outcome', as
 
   await sessionPage.expect.toSeeAppointments()
 
-  const checkAppointmentDetailsPage = await clickUpdateAnAppointment(page, sessionPage, personOnProbation.crn)
+  const checkAppointmentDetailsPage = await viewAppointmentFromList(page, sessionPage, personOnProbation.crn)
   const chooseSupervisorPage = await completeCheckAppointmentDetails(page, checkAppointmentDetailsPage)
 
   const attendanceOutcomePage = await completeChooseSupervisor(page, chooseSupervisorPage, team)
